+++ components/nimbus/src/tests/stateful/test_evaluator.rs
@@ -5,9 +5,12 @@
 use crate::{
     enrollment::NotEnrolledReason,
     evaluator::targeting,
-    stateful::behavior::{
-        EventStore, Interval, IntervalConfig, IntervalData, MultiIntervalCounter,
-        SingleIntervalCounter,
+    stateful::{
+        behavior::{
+            EventStore, Interval, IntervalConfig, IntervalData, MultiIntervalCounter,
+            SingleIntervalCounter,
+        },
+        targeting::RecordedContext,
     },
     tests::helpers::TestRecordedContext,
     AppContext, EnrollmentStatus, TargetingAttributes,
@@ -506,7 +509,7 @@ fn test_multiple_contexts_flatten() -> crate::Result<()> {
     }));
     let mut targeting_attributes =
         crate::tests::test_evaluator::ta_with_locale("en-US".to_string());
-    targeting_attributes.set_recorded_context(&*recorded_context);
+    targeting_attributes.set_recorded_context(recorded_context.to_json());
 
     let value = serde_json::to_value(targeting_attributes).unwrap();